Assign a timeout for client sockets of HAPServer.
Sometimes TCP connections are being lost. For instance, when you move your phone out of a Wi-Fi network, the phone is not able to close the TCP connection with the server and the connection remains open. After a while, an operating system starts refusing to create a new thread: `RuntimeError: can't start new thread`.

This commit introduces a timeout for client sockets. Now, if you move your phone out of a Wi-Fi, a connection between the phone and the server will be closed in a five minutes.